% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func3.R
\name{brandChecker}
\alias{brandChecker}
\title{Womens Shoe Size for Designer Brands}
\usage{
brandChecker(brand = "alexander mcqueen")
}
\arguments{
\item{brand}{The brand, set default to alexander mcqueen}
}
\value{
A shoe size chart
}
\description{
This function allows the user to print out a shoe size conversion chart.
}
\examples{
brandChecker('fendi')
}
\author{
Darius Cuthbert
}
