% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func2.R
\name{brandSize}
\alias{brandSize}
\title{Shoe Size Converter}
\usage{
brandSize(whoShoe = "Woman", brand = "alexander mcqueen", size = 0)
}
\arguments{
\item{whoShoe}{The gender of the user, either male or female, default to female}

\item{brand}{Which sizing standard does the user use, UK or Europe, default to UK}

\item{size}{User's UK or EU shoe size}
}
\value{
User's US shoe size
}
\description{
This function is the main function in this package. It allows users to convert their UK or EU shoe sizes to US sizes.
}
\examples{
brandSize('Woman', 'Fendi', 9)
brandSize('Woman', 'givenchy', 7)
}
\author{
Darius <dcuthbert1@usf.edu>
}
